Split out globals from main into a separate file. No functional changes at this point.
